AHCI: uint32_t vs. size_t issues, hopefully fixes windows buildboxes
svn:sync-xref-src-repo-rev: r110939

+++ src/VBox/Devices/Storage/DevAHCI.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevAHCI.cpp 64021 2016-09-26 17:41:58Z alexander.eichner@oracle.com $ */
+/* $Id: DevAHCI.cpp 64022 2016-09-26 17:44:24Z alexander.eichner@oracle.com $ */
 /** @file
  * DevAHCI - AHCI controller device (disk and cdrom).
  *
@@ -3927,7 +3927,7 @@ static int atapiPassthroughSS(PAHCIREQ pAhciReq, PAHCIPort pAhciPort, size_t cbD
     uint32_t cbTransfer;
     void *pvBuf = NULL;
 
-    cbTransfer = pAhciReq->cbTransfer;
+    cbTransfer = (uint32_t)pAhciReq->cbTransfer;
 
     if (cbTransfer)
     {
